docs:

+++ src/extensions/trash.tsx
@@ -2,6 +2,7 @@ import * as React from 'react'
 import { findChildren } from 'prosemirror-utils'
 import { Extension } from '../types'
 import TrashIcon from '../components/icons/Trash'
+import isMobile from 'is-mobile';
 
 export default class Trash implements Extension {
   get name() {
@@ -47,7 +48,7 @@ export default class Trash implements Extension {
       firstIndex,
       firstIndex + firstNode.node.content.size + 2
     );
-    if (confirm('このブロックを削除してもいいですか？')) {
+    if (isMobile() && confirm('このブロックを削除してもいいですか？')) {
       dispatch(removeTransaction);
     }
   }
